requirements.txt
alembic==1.0.7
aniso8601==4.1.0
Click==7.0
coverage==4.5.2
Flask==1.0.2
Flask-JWT==0.3.2
Flask-Migrate==2.4.0
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.3.2
Flask-Testing==0.7.1
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.1.0
psycopg2-binary==2.7.7
PyJWT==1.4.2
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2018.9
six==1.12.0
SQLAlchemy==1.2.18
Werkzeug==0.14.1
